requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
openai